# Conflicts:
#	app/src/main/java/com/multimoney/multimoney/presentation/navigation/Screen.kt
#	app/src/main/java/com/multimoney/multimoney/presentation/navigation/navgraph/SmartNavGraph.kt
#	app/src/main/java/com/multimoney/multimoney/presentation/ui/home/product/smart/SmartFooterExpanded.kt

+++ app/src/main/java/com/multimoney/multimoney/presentation/ui/credit/disbursement/amount/DisbursementAmountScreen.kt
@@ -144,7 +144,7 @@ fun DisbursementAmountScreen(
                     errorMessage = stringResource(
                         id = viewModel.uiState.disbursementError.second,
                         viewModel.uiState.currencyItems[viewModel.uiState.currencyIndex],
-                        viewModel.uiState.progressFactor
+                        viewModel.uiState.progressFactor.toInt()
                     ),
                     customTransformation = CurrencyIntegerTransformation(
                         viewModel.uiState.currencyItems[viewModel.uiState.currencyIndex],
